Tasks Paralled Execution + DAG + Retry + Executor Pool

+++ src/main/java/com/arsh/workflow/service/impl/WorkflowServiceImpl.java
@@ -16,10 +16,15 @@ import com.arsh.workflow.service.WorkflowExecutorService;
 import com.arsh.workflow.service.WorkflowService;
 import com.arsh.workflow.util.TaskSpecifications;
 import com.arsh.workflow.util.WorkflowSpecifications;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+
+import java.util.concurrent.Executor;
 
 
 @Service
@@ -28,16 +33,30 @@ public class WorkflowServiceImpl implements WorkflowService {
     private final WorkflowRepository workflowRepository;
     private final TaskRepository taskRepository;
     private final WorkflowExecutorService workflowExecutorService;
+    private final Executor executor;
 
-    public WorkflowServiceImpl(WorkflowRepository workflowRepository, TaskRepository taskRepository, WorkflowExecutorService workflowExecutorService) {
+    public WorkflowServiceImpl(WorkflowRepository workflowRepository,
+                               TaskRepository taskRepository,
+                               WorkflowExecutorService workflowExecutorService,
+                               @Qualifier("workflowExecutor") Executor executor
+    ) {
         this.workflowRepository = workflowRepository;
         this.taskRepository = taskRepository;
         this.workflowExecutorService = workflowExecutorService;
+        this.executor = executor;
     }
 
     @Override
     public WorkflowResponse createWorkflow(CreateWorkflowRequest req) {
+
+        String username = SecurityContextHolder.getContext()
+                .getAuthentication()
+                .getName();
+
         Workflow workflow = WorkflowMapper.toEntity(req);
+
+        workflow.setCreatedBy(username);
+        workflow.setUpdatedBy(username);
 
         workflow.setStatus(WorkflowStatus.CREATED);
         workflow = workflowRepository.save(workflow);
@@ -47,10 +66,17 @@ public class WorkflowServiceImpl implements WorkflowService {
 
     @Override
     public WorkflowResponse getWorkflow(Long workflowId) {
+
         Workflow workflow = workflowRepository.findById(workflowId)
                 .orElseThrow(() -> new WorkflowNotFoundException(
                         "Workflow with id " + workflowId + " not found!"
                 ));
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        if (!workflow.getCreatedBy().equals(username)) {
+            throw new AccessDeniedException("You cannot access this workflow");
+        }
 
         return WorkflowMapper.toResponse(workflow);
     }
@@ -58,9 +84,12 @@ public class WorkflowServiceImpl implements WorkflowService {
     @Override
     public WorkflowResponse deleteWorkflow(Long workflowId) {
         Workflow workflow = workflowRepository.findById(workflowId)
-                .orElseThrow(() -> new WorkflowNotFoundException(
-                        "Unable to find workflow with id " + workflowId)
-                );
+                .orElseThrow(() -> new WorkflowNotFoundException("Unable to find workflow with id " + workflowId));
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (!workflow.getCreatedBy().equals(username)) {
+            throw new AccessDeniedException("You cannot delete this workflow");
+        }
 
         WorkflowResponse res = WorkflowMapper.toResponse(workflow);
         workflowRepository.delete(workflow);
@@ -68,13 +97,19 @@ public class WorkflowServiceImpl implements WorkflowService {
         return res;
     }
 
+
     @Override
     public TaskResponse deleteTask(Long workflowId, Long taskId) {
 
         Workflow workflow = workflowRepository.findById(workflowId)
-                .orElseThrow(() -> new WorkflowNotFoundException(
-                        "Workflow with id " + workflowId + " not found"
-                ));
+                .orElseThrow(() -> new WorkflowNotFoundException("Workflow not found"));
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        if (!workflow.getCreatedBy().equals(username)) {
+            throw new AccessDeniedException("Not your workflow");
+        }
+
 
         Task task = workflow.getTasks().stream()
                 .filter(t -> t.getId().equals(taskId))
@@ -95,9 +130,14 @@ public class WorkflowServiceImpl implements WorkflowService {
     public TaskResponse addTask(Long workflowId, CreateTaskRequest req) {
 
         Workflow workflow = workflowRepository.findById(workflowId)
-                .orElseThrow(() -> new WorkflowNotFoundException(
-                        "Workflow with id " + workflowId + "not found.")
-                );
+                .orElseThrow(() -> new WorkflowNotFoundException("Workflow not found"));
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        if (!workflow.getCreatedBy().equals(username)) {
+            throw new AccessDeniedException("Not your workflow");
+        }
+
 
         Task task = TaskMapper.toEntity(req);
 
@@ -117,7 +157,13 @@ public class WorkflowServiceImpl implements WorkflowService {
     public WorkflowResponse startWorkflow(Long workflowId) {
 
         Workflow workflow = workflowRepository.findById(workflowId)
-                .orElseThrow(() ->  new WorkflowNotFoundException("Workflow with id " + workflowId + " not found!"));
+                .orElseThrow(() -> new WorkflowNotFoundException("Workflow not found"));
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        if (!workflow.getCreatedBy().equals(username)) {
+            throw new AccessDeniedException("Not your workflow");
+        }
 
         workflow.getTasks().forEach(t -> {
             if (t.getStatus() == TaskStatus.PENDING) {
@@ -135,10 +181,16 @@ public class WorkflowServiceImpl implements WorkflowService {
 
     @Override
     public WorkflowResponse completeWorkflow(Long workflowId) {
+
         Workflow workflow = workflowRepository.findById(workflowId)
-                .orElseThrow(() -> new WorkflowNotFoundException(
-                        "Workflow with id " + workflowId + " not found."
-                ));
+                .orElseThrow(() -> new WorkflowNotFoundException("Workflow not found"));
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        if (!workflow.getCreatedBy().equals(username)) {
+            throw new AccessDeniedException("Not your workflow");
+        }
+
 
         for (Task task : workflow.getTasks()) {
             if (task.getStatus() != TaskStatus.COMPLETED) {
@@ -162,13 +214,13 @@ public class WorkflowServiceImpl implements WorkflowService {
             String search,
             Pageable pageable
     ) {
-        var spec = WorkflowSpecifications.filter(status, createdBy, search);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        var spec = WorkflowSpecifications.filter(status, username, search);
 
         Page<Workflow> page = workflowRepository.findAll(spec, pageable);
 
-        Page<WorkflowResponse> mapped = page.map(WorkflowMapper::toResponse);
-
-        return PageMapper.toResponse(mapped);
+        return PageMapper.toResponse(page.map(WorkflowMapper::toResponse));
     }
 
     @Override
@@ -178,6 +230,18 @@ public class WorkflowServiceImpl implements WorkflowService {
             String search,
             Pageable pageable
     ) {
+
+        Workflow workflow = workflowRepository.findById(workflowId)
+                .orElseThrow(() -> new WorkflowNotFoundException("Workflow not found"));
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        if (!workflow.getCreatedBy().equals(username)) {
+            throw new AccessDeniedException("You cannot access tasks of another user's workflow");
+        }
+
+
+
         var spec = TaskSpecifications.filter(status, search)
                 .and((root, query, cb) ->
                         cb.equal(root.get("workflow").get("id"), workflowId)
